Cambio de version de springboot: 4.0.1 --> 3.5.8. endpoints login, registro y check email creados.

+++ Streaming-BackEnd/src/main/java/tfg/streamingbackend/entitys/embeddedids/AlbumProductorId.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class AlbumProductorId {
 
     @Column(name = "album_id")
-    private Integer albumId;
+    private Long albumId;
 
     @Column(name = "nombre_productor")
     private String productorNombre;
